pyproject.toml
[project]
name = "ncdb-tools"
version = "0.2.0"
description = "Tools for processing and analyzing National Cancer Database (NCDB) data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jason Brant"},
]
keywords = ["ncdb", "cancer", "database", "analysis", "parquet", "medical-research"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10", 
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Operating System :: OS Independent",
]
dependencies = [
    "polars>=1.0.0",
    "pyarrow>=15.0.0",
    "psutil>=5.9.0",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-psutil>=5.9.0",
    "types-tqdm>=4.65.0",
]
config = [
    "python-dotenv>=1.0.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "RUF"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true

[project.urls]
"Homepage" = "https://github.com/brant01/ncdb-tools"
"Bug Tracker" = "https://github.com/brant01/ncdb-tools/issues"
"Source Code" = "https://github.com/brant01/ncdb-tools"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.12",
    "mypy>=1.16.0",
    "tqdm>=4.67.1",
]
